requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
Django==1.11.12
idna==2.6
kombu==4.1.0
pytz==2018.4
requests==2.18.4
RPi.GPIO==0.6.3
urllib3==1.22
vine==1.1.4
